refactor(#5106): extract placing logic into separate `Placing` class
